+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.4.0 (2018-10-25)
+------------------
+
+* Improved the concurrent creation of agents.
+* Added stop and clear buttons to the interface.
+* Added download button for getting results in excel and json formats.
+* Documentation updated.
+
 0.3.0 (2018-10-01)
 ------------------